backend/backend/requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
faiss-cpu
numpy